+++ xmake/modules/devel/git/ls_remote.lua
@@ -58,7 +58,7 @@ function main(reftype, url)
 
     -- use proxy?
     local envs
-    local proxy_conf = proxy.get(url)
+    local proxy_conf = proxy.config(url)
     if proxy_conf then
         envs = {ALL_PROXY = proxy_conf}
     end